// src/modules/output/interfaces.ts
export interface TreeData {
    label: string;
    nodes?: (TreeData | string)[];
}

export interface TreeOptions {
    prefix?: string
    unicode?: boolean;
}
// ...
export interface ColumnsOptions {
    columns?: string[]
    minWidth?: number
    maxWidth?: number
    align?: 'left' | 'right' | 'center'
    paddingChr?: string
    columnSplitter?: string
    preserveNewLines?: boolean
    showHeaders?: boolean
    dataTransform?: (data) => string
    truncate?: boolean
    truncateMarker?: string
    widths?: { [name: string]: ColumnsOptions }
    config?: { [name: string]: ColumnsOptions }
}


export interface OutputOptions {
    enabled?: boolean
    colors?: boolean
    inspect?: InspectOptions
}